+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## [Unreleased]
-- Support more configurations. Required by Lombok. ([issue #10](https://github.com/coditory/gradle-integration-test-plugin/issues/10))
+
+## [1.3.0] - 2021-08-18
+### Added
+- Add more test based configurations (required by the standard Lombok configuration, [issue #10](https://github.com/coditory/gradle-integration-test-plugin/issues/10))
 
 ## [1.2.3] - 2021-07-20
 ### Changed
@@ -35,7 +38,8 @@
 ### Added
 - Support for intellij idea. Mark integration test directory as a test folder #3.
 
-[Unreleased]: https://github.com/coditory/gradle-integration-test-plugin/compare/v1.2.3...HEAD
+[Unreleased]: https://github.com/coditory/gradle-integration-test-plugin/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/coditory/gradle-integration-test-plugin/compare/v1.2.3...v1.3.0
 [1.2.3]: https://github.com/coditory/gradle-integration-test-plugin/compare/v1.2.2...v1.2.3
 [1.2.2]: https://github.com/coditory/gradle-integration-test-plugin/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/coditory/gradle-integration-test-plugin/compare/v1.2.0...v1.2.1